Does anyone know a neat way to get mouse position on Android?
The crazy way is to create a topmost global view and capture
mouse events on it.

Android has mouse acceleration, which means we cannot sum all
relative moves and conclude the final mouse position, i.e.
mouseX and mouseY here are inaccurate. The user cannot move
the mouse sometimes (i.e. our code believes mouseX is 0 cannot
move any lefter but mouseX should be 421 etc.)

This patch workarounds the annoying issue, by continuing to
send move events when the cursor is moving at screen edges.

The patch works best with android screen on the right of some
other screen. Edit clearMousePosition if not.

The original code uses mixed TAB and SPACE. I'm not cleaning
it up in this commit.